Update DoubleLinkedList.rb
Compare this file with https://github.com/miriam-cortes/CSFunSoloPractice/blob/master/DoubleLinkedList.rb
See comments inline starting with "Shruti:". There is a bug in method addNodeToBeginning. Unless the requirements specifically ask to add a node to the end of the linked list, always default to add node to the beginning since it is more efficient. See also the comment added to share the difference between question #2 and question #4.